allow different grammars for different file types
